+++ src/org/schema/ProgramMembership.php
@@ -10,15 +10,15 @@ class ProgramMembership extends Intangible
 {
     /**
      * The Organization (airline, travelers' club, retailer, etc.) the membership is made with or which offers the MemberProgram.
-     * @var null|Organization
+     * @var null|array|Organization
      */
-    public null|Organization $hostingOrganization ;
+    public null|array|Organization $hostingOrganization ;
 
     /**
      * A member of an Organization or a ProgramMembership. Organizations can be members of organizations; ProgramMembership is typically for individuals.
-     * @var Organization|Person|null
+     * @var Organization|Person|array|null
      */
-    public null|Organization|Person $member ;
+    public null|array|Organization|Person $member ;
 
     /**
      * A unique identifier for the membership.
@@ -29,19 +29,19 @@ class ProgramMembership extends Intangible
     /**
      * The number of membership points earned by the member.
      * If necessary, the unitText can be used to express the units the points are issued in. (E.g. stars, miles, etc.)
-     * @var int|QuantitativeValue|null
+     * @var int|QuantitativeValue|array|null
      */
-    public null|int|QuantitativeValue $membershipPointsEarned ;
+    public null|int|array|QuantitativeValue $membershipPointsEarned ;
 
     /**
      * The MemberProgram associated with a ProgramMembership.
-     * @var ?MemberProgram
+     * @var null|array|MemberProgram
      */
-    public ?MemberProgram $program ;
+    public null|array|MemberProgram $program ;
 
     /**
      * The program providing the membership. It is preferable to use :program instead.
-     * @var ?string
+     * @var string|null
      */
     public ?string $programName ;
 }