+++ src/org/schema/constants/traits/ProgramMembership.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace org\schema\constants\traits;
+
+/**
+ * The ProgramMembership properties enumeration.
+ */
+trait ProgramMembership
+{
+    const string HOSTING_ORGANIZATION     = 'hostingOrganization' ;
+    const string MEMBER                   = 'member' ;
+    const string MEMBERSHIP_NUMBER        = 'membershipNumber' ;
+    const string MEMBERSHIP_POINTS_EARNED = 'membershipPointsEarned' ;
+    const string PROGRAM                  = 'program' ;
+    const string PROGRAM_NAME             = 'programName' ;
+}