+++ src/org/schema/constants/traits/Properties.php
@@ -70,6 +70,7 @@ trait Properties
         Product ,
         ProductGroup ,
         ProductModel ,
+        ProgramMembership ,
         Rating ,
         Review ,
         Service ,